// angular-poc/src/types/datalayer/airindia.d.ts
export interface AirIndiaDL {
  isErrorPage: boolean;
  siteSection: string;
  siteSubSection: string;
  countryCode?: string;
  countryName?: string;
  language?: string;
  pageName?: string;
  pageType?: string;
  serverName?: string;
  siteName?: string;
  linkURL?: string;
  pageURL?: string;
  userDepartment?: string;
  userDesignation?: string;
  userFunction?: string;
  userPersona?: string;
  userEmploymentType? :string;
}